Start simplifying job to its core component.
It is essentially just a handle to monitor a group of tasks. It's barely more than a glorified shared_ptr to state.
The current set of changes aren't quite minimal enough, and open us up to a race condition. The general
interface for spawning a set of tasks into a job is mostly the problem; there shouldn't be a way to add tasks
to a completed job, but only to create a job with a set of initial tasks and a way to add child tasks from within
a task.