Added the ability to park a thread on a queue (wait for jobs).
The general approach is ideally efficient - no semaphores or other thread primitives need to be signaled if there are no parked threads.
The specific implementation is a little heavier than desired (like our underlying queue), but notable the current code has a race
condition. This race affects performance but not correctness; namely, if two tasks are spawned and two threads are parked, it may
happen that only the first thread is unparked.