Replace mutex and deque from parking_lot with a (oversimple) spinlock and linked list.
The idea is for parking_lot to be embedded in job, so it needs to be very lightweight to construct even if unused and shouldn't
require allocations for a single use, which made both the old mutex and the deque poor fits for performance reasons.

The linked list in there now is a very simple singly-linked list that functions as a super basic queue. This is going to need to
be extended to deal with parkables that park in multiple lots, since it'll be necessary for a parkable to be removed from a
lot even if it's not at the head of the lot's queue.